+++ src/components/shared/OrderSummary/OrderSummary.jsx
@@ -1,15 +1,8 @@
 import React from "react";
-import Invoice from "../Invoice/Invoice";
 import { Box, Divider, Typography } from "@mui/material";
 
-const OrderDetails = [
-  { label: "Sub Total", amount: 119.69 },
-  { label: "Discount", amount: -13.4 },
-  { label: "Delivery Fee", amount: 0.0 },
-  { label: "Grand Total", amount: 106.29 },
-];
 
-export default function FinancialBreakdownList({ title, children }) {
+export default function OrderSummary({ title, children }) {
   return (
     <Box sx={{ maxWidth: 360 }}>
       <Typography
@@ -27,7 +20,6 @@ export default function FinancialBreakdownList({ title, children }) {
         {title}
       </Typography>
       <Divider width={"100%"} />
-      <Invoice OrderDetails={OrderDetails} />
       {children}
     </Box>
   );
